Update minor changes to adhere to PEP8 guidelines
